requirements.txt
# Core
python>=3.9

# Machine Learning
xgboost>=1.7
scikit-learn>=1.3
numpy
pandas
joblib

# Web App
streamlit

# Feature Extraction
requests
beautifulsoup4
python-whois
dnspython

# Utilities
urllib3
